Added Buffer class
Also added small template class for a renderer. Still need to see how to make a nice interface for that.
We are now passing a pointer to the owning application of a layer in OnAttach. This way the layer has a way to access the window (to show FPS in title bar).